requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2024.2.2
click==8.1.7
fastapi==0.109.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.7
pip==23.3.2
pydantic-core==2.16.1
pydantic==2.6.0
python-dotenv==1.0.0
setuptools==69.0.3
sniffio==1.3.0
starlette==0.36.3
typing-extensions==4.9.0
uvicorn==0.27.0
wheel==0.42.0
